# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ТОПливо — Fuel Voucher API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TelegramAuthInput {
  telegramId: number;
  firstName: string;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  photoUrl?: string | null;
}

export interface User {
  id: number;
  telegramId: number;
  firstName: string;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  isAdmin?: boolean;
  totalVouchers?: number;
  totalSavings?: number;
  createdAt: string;
}

export interface FuelPrice {
  /** AI-92 | AI-95 | AI-98 | Diesel */
  fuelType: string;
  pricePerLiter: number;
  /** @nullable */
  lockedPrice?: number | null;
  updatedAt: string;
}

export interface Station {
  id: number;
  name: string;
  /** Lukoil | Gazprom | Rosneft | Shell | Tatneft */
  network: string;
  address: string;
  lat: number;
  lng: number;
  isActive: boolean;
  networkColor?: string;
  prices?: FuelPrice[];
}

export interface LivePrices {
  ai92?: number;
  ai95?: number;
  ai98?: number;
  diesel?: number;
  updatedAt: string;
  /** up | down | stable */
  trend?: string;
}

export interface PriceHistoryPoint {
  date: string;
  fuelType: string;
  price: number;
}

export interface Voucher {
  id: number;
  userId: number;
  stationId: number;
  stationName?: string;
  stationNetwork?: string;
  fuelType: string;
  liters: number;
  pricePerLiter: number;
  totalAmount: number;
  lockedAt: string;
  expiresAt: string;
  /** active | used | expired */
  status: string;
  /** @nullable */
  qrCode?: string | null;
  /** @nullable */
  savingsAmount?: number | null;
  /** @nullable */
  paymentMethod?: string | null;
}

export interface VoucherInput {
  stationId: number;
  fuelType: string;
  liters: number;
  /** stars | crypto */
  paymentMethod: string;
  telegramUserId?: number;
}

export interface AnalyticsSummary {
  totalVouchers: number;
  totalLiters: number;
  totalSavings: number;
  activeVouchers: number;
  avgSavingsPercent?: number;
  marketPriceAi95?: number;
  lockedPriceAi95?: number;
  priceGrowthSinceLock?: number;
}

export interface SupplyMatrixRow {
  network: string;
  /** available | low | unavailable */
  ai92: string;
  ai95: string;
  ai98: string;
  diesel: string;
  totalStations: number;
  networkColor?: string;
}

export interface MarketComparison {
  fuelType: string;
  liters: number;
  lockedPrice: number;
  marketPrice: number;
  lockedTotal?: number;
  marketTotal?: number;
  savings: number;
  savingsPercent: number;
}

export interface AdminStats {
  totalUsers: number;
  totalVouchers: number;
  activeVouchers: number;
  usedVouchers?: number;
  expiredVouchers?: number;
  totalRevenue: number;
  totalLitersSold?: number;
  newUsersToday?: number;
  vouchersToday?: number;
}

export interface BroadcastInput {
  message: string;
  adminSecret?: string;
}

export interface BroadcastResult {
  sent: number;
  failed: number;
}

export type ListStationsParams = {
network?: string;
fuelType?: string;
};

export type GetPriceHistoryParams = {
fuelType?: string;
days?: number;
};

export type GetMarketComparisonParams = {
fuelType?: string;
liters?: number;
};

export type ListAdminUsersParams = {
limit?: number;
offset?: number;
};