# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ТОПливо — Fuel Voucher API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  totalUsers: number;
  totalVouchers: number;
  activeVouchers: number;
  usedVouchers?: number;
  expiredVouchers?: number;
  totalRevenue: number;
  totalLitersSold?: number;
  newUsersToday?: number;
  vouchersToday?: number;
}