# lib/api-zod/src/generated/types/analyticsSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ТОПливо — Fuel Voucher API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyticsSummary {
  totalVouchers: number;
  totalLiters: number;
  totalSavings: number;
  activeVouchers: number;
  avgSavingsPercent?: number;
  marketPriceAi95?: number;
  lockedPriceAi95?: number;
  priceGrowthSinceLock?: number;
}