# lib/api-zod/src/generated/types/broadcastInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ТОПливо — Fuel Voucher API
 * OpenAPI spec version: 0.1.0
 */

export interface BroadcastInput {
  message: string;
  adminSecret?: string;
}