# lib/api-zod/src/generated/types/fuelPrice.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ТОПливо — Fuel Voucher API
 * OpenAPI spec version: 0.1.0
 */

export interface FuelPrice {
  /** AI-92 | AI-95 | AI-98 | Diesel */
  fuelType: string;
  pricePerLiter: number;
  /** @nullable */
  lockedPrice?: number | null;
  updatedAt: string;
}