# lib/api-zod/src/generated/types/station.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ТОПливо — Fuel Voucher API
 * OpenAPI spec version: 0.1.0
 */
import type { FuelPrice } from './fuelPrice';

export interface Station {
  id: number;
  name: string;
  /** Lukoil | Gazprom | Rosneft | Shell | Tatneft */
  network: string;
  address: string;
  lat: number;
  lng: number;
  isActive: boolean;
  networkColor?: string;
  prices?: FuelPrice[];
}