# lib/api-zod/src/generated/types/supplyMatrixRow.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ТОПливо — Fuel Voucher API
 * OpenAPI spec version: 0.1.0
 */

export interface SupplyMatrixRow {
  network: string;
  /** available | low | unavailable */
  ai92: string;
  ai95: string;
  ai98: string;
  diesel: string;
  totalStations: number;
  networkColor?: string;
}