# lib/api-zod/src/generated/types/voucher.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ТОПливо — Fuel Voucher API
 * OpenAPI spec version: 0.1.0
 */

export interface Voucher {
  id: number;
  userId: number;
  stationId: number;
  stationName?: string;
  stationNetwork?: string;
  fuelType: string;
  liters: number;
  pricePerLiter: number;
  totalAmount: number;
  lockedAt: string;
  expiresAt: string;
  /** active | used | expired */
  status: string;
  /** @nullable */
  qrCode?: string | null;
  /** @nullable */
  savingsAmount?: number | null;
  /** @nullable */
  paymentMethod?: string | null;
}